Move feature to contrib package
